requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
django-import-export==2.8.0
django-ninja==0.19.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.3
pydantic==1.9.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.3.0
tzdata==2022.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
